+++ src/edna_processor/hap_network_nex_generator.py
@@ -4,8 +4,8 @@ import numpy as np
 import os
 import pandas as pd
 import tempfile
-from edna_processor.utils_hdbscan import fit_hdbscan
-from edna_processor.utils_sequence import write_fasta, align_fasta, get_uniq_seq_freq
+from edna_processor.utils.utils_hdbscan import fit_hdbscan
+from edna_processor.utils.utils_sequence import write_fasta, align_fasta, get_uniq_seq_freq
 
 class HapNetNexusGenerator(AnalysisManager):
